# agrieasy/docs/openapi.yaml
openapi: 3.0.3
info:
  title: FarmCare API
  version: 1.0.0
servers:
  - url: http://localhost:3001
paths:
  /auth/request-otp:
    post:
      summary: Request OTP (mock)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
              required: [phone]
      responses:
        '200':
          description: OTP issued (mocked)
  /auth/verify-otp:
    post:
      summary: Verify OTP and get JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                otp:
                  type: string
                name:
                  type: string
                role:
                  type: string
                  enum: [FARMER,EQUIPMENT_VENDOR,INPUT_SUPPLIER,TRANSPORTER,CONSUMER,ADMIN]
              required: [phone, otp]
      responses:
        '200':
          description: JWT token
...
  /equipment:
    get:
      summary: List equipment
    post:
      summary: Create equipment (vendor)
      security:
        - bearerAuth: []
...
  /produce:
    get:
      summary: List produce
    post:
      summary: Create produce (farmer)
      security:
        - bearerAuth: []
...
  /orders/checkout:
    post:
      summary: Checkout cart to create order
      security:
        - bearerAuth: []
  /orders/{id}:
    get:
      summary: Get order by id
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT